Fixed UI issues
